% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalbasic.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion (BIC) value.}
\usage{
bic(ttnsr,theta,omega,d,r)
}
\arguments{
\item{ttnsr}{An observed tensor.}

\item{theta}{A continuous-valued tensor (latent parameters).}

\item{omega}{The cut-off points.}

\item{d}{Dimension of the tensor.}

\item{r}{Rank of the tensor.}
}
\value{
BIC value at given inputs based on cumulative logistic model.
}
\description{
Compute Bayesian Information Criterion (BIC) given a parameter tensor, an observed tensor, the dimension, and the rank based on cumulative logistic model. This BIC function is designed for selecting rank in the \code{fit_ordinal} function.
}
