% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalbasic.R
\name{likelihood}
\alias{likelihood}
\title{Log-likelihood function (cost function).}
\usage{
likelihood(ttnsr,theta,omega,type = c("ordinal","Gaussian"))
}
\arguments{
\item{ttnsr}{An observed tensor data.}

\item{theta}{A continuous-valued tensor (latent parameters).}

\item{omega}{The cut-off points.}

\item{type}{Types of log-likelihood function.

\code{"ordinal"} specifies log-likelihood function based on the cumulative logistic model.

\code{"Gaussian"} specifies log-likelihood function based on the Gaussian model.}
}
\value{
Log-likelihood value at given inputs.
}
\description{
Return log-likelihood function (cost function) value evaluated at a given parameter tensor, an observed tensor, and cut-off points.
}
