% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalbasic.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion (BIC) value.}
\usage{
bic(ttnsr,theta,omega,d,r)
}
\arguments{
\item{ttnsr}{an observed tensor.}

\item{theta}{a continuous-valued tensor (latent parameters).}

\item{omega}{the cut-off points.}

\item{d}{dimension of the tensor.}

\item{r}{rank of the tensor.}
}
\value{
BIC value at given inputs.
}
\description{
Compute Bayesian Information Criterion (BIC) given a parameter tensor, an observed tensor, the dimension, and the rank.
}
